requirements.txt
Django==1.9.2
psycopg2==2.6.1
django-ckeditor==5.0.3
pytz==2015.7
django-bower==5.1.0
nodeenv==0.13.6
django-import-export==0.4.2
django-leaflet==0.18.0
django-crispy-forms==1.6.0
django-registration-redux==1.4
django-related-admin==0.3.0
django-autocomplete-light==3.1.1
